refactored name of llm middleware
