+++ packages/railtracks/src/railtracks/built_nodes/easy_usage_wrappers/agent.py
@@ -7,7 +7,7 @@ from railtracks.built_nodes.concrete import (
     RTFunction,
 )
 from railtracks.built_nodes.concrete.response import StringResponse, StructuredResponse
-from railtracks.built_nodes.middlewares.core import ModelMiddleware
+from railtracks.built_nodes.llm.middleware.core import ModelMiddleware
 from railtracks.guardrails.core import Guard
 from railtracks.llm.message import SystemMessage
 from railtracks.llm.tools.parameters._base import Parameter
